requirements.txt
aiofiles==23.1.0
asgiref==3.6.0
colorama==0.4.6
dict-toolbox==3.1.0
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
msgpack==1.0.5
nest-asyncio==1.5.6
pop-config==12.0.2
pop-loop==1.0.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
PyYAML==6.0
rend==6.5.2
sniffio==1.3.0
sqlparse==0.4.4
toml==0.10.2
whitenoise==6.4.0
